also capture *.jpg

+++ download_img.rb
@@ -2,10 +2,10 @@
 require "open-uri"
 require "json"
 
-pat1 = /<img class="media-object" src="(https:\/\/www.keepon.com.tw\/UploadFile\/\w+\/\w+\/\w+\/([\w\-]+\.png))" width="100" height="100" alt="image">/
+pat1 = /<img class="media-object" src="(https:\/\/www.keepon.com.tw\/UploadFile\/\w+\/\w+\/\w+\/([\w\-]+\.[jpn]{2}g))" width="100" height="100" alt="image">/
 pat2 = /https:\/\/www.keepon.com.tw\/UploadFile\/\w+\/\w+\/\w+\/([\w\-]+\.png)/
 
-html = open("index.html").read
+html = open("raw.html").read
 urls = html.scan(pat1).to_a
 
 unless Dir.exists?("images/")
@@ -16,9 +16,10 @@ end
 for url in urls
 	puts "fetching " + url[0] 
 	filename = url[1]
-	open("images/" + filename, 'wb') do |file|
-		puts "saving " + filename
-        file << open(url[0].to_s).read
-    end
-
+	unless File.exists?("images/" + filename)
+		open("images/" + filename, 'wb') do |file|
+			puts "saving " + filename
+	        file << open(url[0].to_s).read
+	    end
+	end
 end
